prevent auto deletion of system vars and factors, by deleting back refs
(cherry picked from commit 7dbe4a1)

+++ globalVar.js
@@ -256,7 +256,7 @@ GlobalVar.prototype.removeBackRef = function(entity) {
 
     // check if variable is now completely without backRefs, then remove from entity list (update)
     if (window.uc!==undefined) {
-        if (backRefs.length == 0) {
+        if (backRefs.length == 0 && !this.isFactor() && !this.expData.isSystemVar(this)) {
             this.expData.deleteGlobalVar(this);
         }
     }
